use uint8

+++ pkg/bpf/probe/probe.go
@@ -58,7 +58,7 @@ func (p *Probe) UpdateMap(ips []string) {
 		}
 		ipBytes := binary.BigEndian.Uint32(ip4)
 
-		err := p.bpfObjs.BannedIps.Put(ipBytes, 0)
+		err := p.bpfObjs.BannedIps.Put(ipBytes, uint8(0))
 		if err != nil {
 			// log it
 		}
